requirements.txt
Django
django-cors-headers
django-environ
django-jalali-date
django-middleware-global-request
django-taggit
django-user-agents
django-utils-six
djangorestframework
djangorestframework-simplejwt
filelock
greenlet
idna
jalali_core
jdatetime
legacy-cgi
mssql-django
platformdirs
pyasn1
pyasn1-modules
pycparser
pycryptodome
PyJWT
pyodbc
python-dateutil
python-dotenv
pytz
requests
six
SQLAlchemy
sqlparse
typing_extensions
tzdata
ua-parser
ua-parser-builtins
urllib3
user-agents
Faker
fastapi
pydantic
